fix: run nomad jobs from folder

+++ runall.sh
@@ -1,13 +1,12 @@
 #!/bin/bash
 set -euo pipefail
 
+[ -d "${1:-}" ] || exit 1
+
 docker-compose down
 docker-compose up -d
 curl --request PUT --data @external-services/postgres.json 172.16.1.101:8500/v1/catalog/register
 
-nomad job run jobs/cli-jobs/petclinic-web.nomad
-nomad job run jobs/cli-jobs/petclinic-egw.nomad
-nomad job run jobs/cli-jobs/petclinic-api.nomad
-nomad job run jobs/cli-jobs/petclinic-ingw.nomad
+find "$1" -name '*.nomad' -type f -exec nomad job run -detach {} \;
 
-for i in consul-config/intention-*; do consul config write $i; done
+#for i in consul-config/intention-*; do consul config write $i; done
